Added CoDrone.Delay(int interval)
Added Functions:
- There is a known issue where the CoDrone will not run any command for longer 
than ~1 second. This is likely due to an internal timeout function on the CoDrone
firmware. Because of this, using delay() for basic autonomous functions does not work. 
Ideally, the following code would send the CoDrone flying upwards for 5 seconds:
      THROTTLE = 100;
      CoDrone.Control();
      delay(5000);
but the timeout issue would make this CoDrone only fly upwards for ~1 second.
To circumvent this, a new CoDrone.Delay function was created.

 + CoDrone.Delay(int interval);
   - Repeatedly sends the last throttle, roll, pitch and yaw commands once every
   500 milliseconds for the duration of the interval
   - Uses a while loop to imitate delay
   - New class variables Prevroll, Prevyaw, Prevpitch, Prevthrottle