Moved TimeCheck3() to BattleReceive()
Putting a TimeCheck in Receive() was a pretty risky move.  It ended up
not paying off, and instead messed up a lot of other things.

By putting the check in BattleReceive(), the time check only occurs
specifically for the battle function.  This will no longer interfer with
send_check or Connect.

Additionally, BATTLE_CHECK_TIME has been increased from 5 to 10