Create Controller_Flight_Custom.ino
NOTICE: I am very, very new to Arduino but experienced in Lua; this is how we can define tables in Lua,  by using table["ItemInTable"] = 2; etc. Please tell me if this is not in .ino! 
Also please do some debugging, I'll try to run it on my CoDrone as well. :)
Thanks! - Nabfam